requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.12.0
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.9.1
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.2.6
packaging==25.0
parso==0.8.5
platformdirs==4.5.0
prompt_toolkit==3.0.52
psutil==7.1.3
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.2.14
faiss-cpu==1.7.4          # FAISS library for similarity search (CPU version)
requests==2.31.0           # HTTP requests
tqdm==4.66.1               # Progress bars
pypdf==3.15.1              # PDF reading
sentence-transformers==2.2.2 # SentenceTransformer embeddings
python-dotenv==1.0.1       # Load environment variables from .env file